pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ProtMiner"
version = "0.5.0"
description = "ProtMiner: inclusive protein family discovery using HMMER, DIAMOND, InterPro, motifs, with optional topology/signal filters."
authors = [{name = "Baradwaj Ravi Gopal"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.2",
    "biopython>=1.79",
    "argcomplete>=3.0.0"
]

[project.scripts]
find-proteins = "protscout.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
